powerbuilder/script/requirements.txt
aiohappyeyeballs==2.6.1
alpaca-py==0.42.1
alpaca-trade-api==3.2.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
aiohttp==3.12.15
aiosignal==1.4.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
attrs==25.3.0
auto-py-to-exe==2.46.0
beautifulsoup4==4.13.4
blinker==1.9.0
bottle==0.13.4
bottle-websocket==0.2.9
cabarchive==0.2.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
configparser==7.2.0
contourpy==1.3.3
curl_cffi==0.13.0
cx_Freeze==8.3.0
cx_Logging==3.2.1
cycler==0.12.1
distro==1.9.0
Eel==0.18.2
eventkit==1.0.3
feedparser==6.0.11
filelock==3.18.0
Flask==3.1.1
Flask-MySQLdb==2.0.0
fonttools==4.59.0
frozendict==2.4.6
frozenlist==1.7.0
future==1.0.0
gevent==25.5.1
gevent-websocket==0.10.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
ib-insync==0.9.86
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
kiwisolver==1.4.8
lief==0.16.5
linecache2==1.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mplfinance==0.12.10b0
multidict==6.6.3
multitasking==0.0.12
mysql-connector==2.2.9
mysql-connector-python==9.4.0
mysqlclient==2.2.7
nest-asyncio==1.6.0
nltk==3.9.1
notifications==0.3.2
numpy==2.3.2
openai==0.28.0
packaging==25.0
pandas==2.3.1
pathlib==1.0.1
peewee==3.18.2
pefile==2023.2.7
pillow==11.3.0
pip==25.2
platformdirs==4.3.8
propcache==0.3.2
protobuf==6.31.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
PyMySQL==1.1.1
pyodbc==5.2.0
pyparsing==3.2.3
pypiwin32==223
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
regex==2025.7.34
requests==2.32.4
schedule==1.2.2
scipy==1.16.1
setuptools==80.9.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
striprtf==0.0.29
tabulate==0.9.0
textblob==0.19.0
tqdm==4.67.1
traceback2==1.4.0
typing_extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
win10toast==0.9
yarl==1.20.1
yfinance==0.2.65
zope.event==5.1.1
zope.interface==7.2
